+++ CHANGELOG.md
@@ -10,6 +10,11 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- General
+
+  - All packages can now be used with Java 11
+    ([#45](https://github.com/utybo/Tegral/pull/45))
+
 - `tegral-openapi-ktor-resources`
 
   - Initial release ([#43](https://github.com/utybo/Tegral/pull/43))
@@ -204,9 +209,6 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - General
-
-  - All packages can now be used with Java 11
-    ([#45](https://github.com/utybo/Tegral/pull/45))
 
   - Bumped Ktor version to version 2.0.3
     ([#17](https://github/utybo/Tegral/pull/17))